# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Streambox API
  description: API REST para la gestión de películas, géneros y usuarios de Streambox.
  version: '1.0'
servers:
- url: http://localhost:8080
  description: Backend local (perfil dev)
paths:
...
  /api/auth/refresh:
    post:
      tags:
      - auth-controller
      summary: Renueva la sesión
      description: 'Entrega un JWT de acceso nuevo (cookie streambox_token) a partir del refresh token de la cookie streambox_refresh,
        sin pedir la contraseña. Sin cuerpo. Solo lee la cookie streambox_refresh: ignora streambox_token y la cabecera Authorization.
        Exige la cabecera X-Requested-With: StreamBox (defensa CSRF). Cada uso rota el refresh token: el viejo queda revocado
        y llega uno nuevo en Set-Cookie (vida de 7 días, sin pasar de 30 días desde el login). Reutilizar un refresh token
        ya rotado se trata como un robo y revoca la sesión entera (también el token nuevo), salvo en los 10 segundos siguientes
        a la rotación: así dos pestañas que renuevan a la vez no se cierran la sesión; la segunda recibe solo el JWT nuevo,
        sin Set-Cookie de streambox_refresh, porque el navegador ya tiene el sucesor.'
      operationId: refresh
      parameters:
      - name: streambox_refresh
        in: cookie
        description: Refresh token (cookie HttpOnly streambox_refresh que fija el login; el navegador la envía solo)
        required: false
        schema:
          type: string
      responses:
        '204':
          description: Sesión renovada; sin cuerpo. Set-Cookie con streambox_token y, salvo en la gracia de un refresh simultáneo,
            con el streambox_refresh nuevo
        '401':
          description: 'Código SESSION_EXPIRED: el refresh token falta, no existe, ha caducado (él o su sesión de 30 días),
            está revocado (logout) o se ha reutilizado (y entonces se revoca toda la sesión). Mismo cuerpo en todos los casos.
            La respuesta borra las dos cookies: hay que volver a iniciar sesión'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Código CSRF_REJECTED: falta la cabecera X-Requested-With: StreamBox. No se mira el refresh token ni
            se gasta el límite por IP'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 'Código RATE_LIMIT_EXCEEDED: demasiadas renovaciones desde esta IP (30 por minuto por defecto); la
            cabecera Retry-After indica los segundos que hay que esperar. Solo cuentan las peticiones con la cabecera X-Requested-With
            y con la cookie streambox_refresh: sin cookie (visita sin sesión) la respuesta es siempre el 401 SESSION_EXPIRED'
          headers:
            Retry-After:
              description: Segundos que hay que esperar antes de volver a intentarlo (entero, mínimo 1).
              required: true
              schema:
                type: integer
                format: int32
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: 'Formato común de todas las respuestas de error de la API (4xx y 5xx). El cliente debe decidir por status
        y code, nunca por el texto de message. Los campos que no aplican a un error se omiten del JSON: nunca se envían con
        valor null.'
      properties:
        timestamp:
          type: string
          format: date-time
          description: Fecha y hora del error, en UTC y formato ISO-8601.
          example: '2026-10-04T11:15:30.123Z'
        status:
          type: integer
          format: int32
          description: Código de estado HTTP; coincide con el de la respuesta.
          example: 404
        error:
          type: string
          description: Frase estándar del estado HTTP según la especificación (en inglés).
          example: Not Found
        code:
          type: string
          description: 'Código funcional estable del error: el cliente debe decidir por este campo (y por status), nunca por
            el texto de message. Valores:

            - RESOURCE_NOT_FOUND (404): el recurso o la ruta no existen (una serie sin episodios también da este código a
            los usuarios).

            - MOVIE_NOT_IN_FAVORITES (404): la película no está en la lista del usuario.

            - SERIES_NOT_IN_FAVORITES (404): la serie existe, pero no está en la lista del usuario.

            - CSRF_REJECTED (403): petición no segura (POST/PUT/PATCH/DELETE) autenticada por cookie sin la cabecera X-Requested-With:
            StreamBox. POST /api/auth/refresh y POST /api/auth/logout la exigen siempre, también con Authorization: Bearer.

            - USER_ALREADY_EXISTS (409): el nombre de usuario o el email ya están registrados.

            - MOVIE_ALREADY_IN_FAVORITES (409): la película ya está en la lista del usuario.

            - SERIES_ALREADY_IN_FAVORITES (409): la serie ya está en la lista del usuario.

            - EPISODE_ALREADY_EXISTS (409): la serie ya tiene un episodio con esa temporada y número.

            - GENRE_ALREADY_EXISTS (409): ya existe otro género con ese nombre.

            - GENRE_IN_USE (409): el género no se puede borrar porque alguna película o serie lo tiene asignado.

            - AMBIGUOUS_TITLE (409): varias películas tienen ese título; hay que usar el id.

            - DATA_INTEGRITY_VIOLATION (409): la operación choca con una restricción de los datos.

            - VALIDATION_ERROR (400): datos o parámetros no válidos; el detalle por campo va en validationErrors.

            - CURRENT_PASSWORD_INCORRECT (400): al cambiar la contraseña, la actual no es correcta; el detalle va en validationErrors.currentPassword.
            No es un 401: la sesión sigue siendo válida.

            - MALFORMED_REQUEST (400 y otros 4xx genéricos): la petición no se puede interpretar (JSON mal formado, tipos
            incorrectos).

            - INVALID_CREDENTIALS (401): email o contraseña incorrectos en el login, o token ausente, caducado o no válido.

            - SESSION_EXPIRED (401): POST /api/auth/refresh no puede renovar la sesión (refresh token ausente, caducado, revocado
            o reutilizado); hay que volver a iniciar sesión.

            - ACCESS_DENIED (403): autenticado, pero sin permisos (por ejemplo, no es ADMIN).

            - METHOD_NOT_ALLOWED (405): el método HTTP no está soportado por la ruta.

            - UNSUPPORTED_MEDIA_TYPE (415): el tipo de contenido no está soportado.

            - NOT_ACCEPTABLE (406): la cabecera Accept pide un formato que la API no produce; la API solo responde en JSON
            (este error también va en JSON).

            - RATE_LIMIT_EXCEEDED (429): demasiadas peticiones de login, registro o renovación de sesión (POST /api/auth/refresh)
            desde la misma IP, o demasiados intentos con la contraseña actual incorrecta en PUT /api/users/me/password (por
            cuenta: 5 cada 15 minutos); ver la cabecera Retry-After.

            - ACCOUNT_LOCKED (429): la cuenta está bloqueada temporalmente por logins fallidos; ver la cabecera Retry-After.

            - INTERNAL_ERROR (500 u otro 5xx): error no controlado en el servidor.'
          enum:
          - RESOURCE_NOT_FOUND
          - USER_ALREADY_EXISTS
          - MOVIE_ALREADY_IN_FAVORITES
          - MOVIE_NOT_IN_FAVORITES
          - VALIDATION_ERROR
          - INVALID_CREDENTIALS
          - DATA_INTEGRITY_VIOLATION
          - INTERNAL_ERROR
          - AMBIGUOUS_TITLE
          - ACCESS_DENIED
          - MALFORMED_REQUEST
          - METHOD_NOT_ALLOWED
          - UNSUPPORTED_MEDIA_TYPE
          - NOT_ACCEPTABLE
          - RATE_LIMIT_EXCEEDED
          - ACCOUNT_LOCKED
          - GENRE_ALREADY_EXISTS
          - GENRE_IN_USE
          - EPISODE_ALREADY_EXISTS
          - SERIES_ALREADY_IN_FAVORITES
          - SERIES_NOT_IN_FAVORITES
          - CSRF_REJECTED
          - SESSION_EXPIRED
          - CURRENT_PASSWORD_INCORRECT
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          description: 'Mensaje legible en español, pensado para mostrarlo al usuario. Puede cambiar: no se debe usar para
            tomar decisiones.'
          example: Película no encontrada
        path:
          type: string
          description: Ruta de la petición que produjo el error (sin los parámetros de consulta).
          example: /api/movies/42
        validationErrors:
          type: object
          additionalProperties:
            type: string
          description: 'Solo en los 400 con código VALIDATION_ERROR o CURRENT_PASSWORD_INCORRECT: un mensaje por cada campo
            o parámetro no válido (nombre del campo → mensaje). En el resto de errores no aparece (no se envía con null).'
          example:
            title: El título es obligatorio
        remainingAttempts:
          type: integer
          format: int32
          description: 'Solo en el 401 INVALID_CREDENTIALS del login: intentos que le quedan a la cuenta, con este fallo ya
            descontado. Nunca vale 0, porque el fallo que agota los intentos ya responde 429 ACCOUNT_LOCKED. En el resto de
            errores no aparece (no se envía con null).'
          example: 4
          minimum: 1
      required:
      - code
      - error
      - message
      - path
      - status
      - timestamp
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: 'Autenticación principal (navegador): el JWT de acceso viaja en la cookie HttpOnly streambox_token que
        fija POST /api/auth/login (Path=/api, 15 minutos; se renueva con POST /api/auth/refresh). Las peticiones no seguras
        (POST, PUT, PATCH, DELETE) autenticadas por esta cookie deben llevar además la cabecera X-Requested-With: StreamBox
        (defensa CSRF); si falta, 403 CSRF_REJECTED. Si la petición trae Authorization: Bearer, manda el Bearer y la cookie
        no se mira. Swagger UI no puede fijar esta cookie: para probar desde aquí usa bearerAuth.'
      name: streambox_token
      in: cookie
    bearerAuth:
      type: http
      description: 'Alternativa para clientes de API sin navegador: el mismo JWT de acceso (el valor de la cookie streambox_token)
        en la cabecera Authorization: Bearer. Con Bearer no se exige X-Requested-With (no hay riesgo de CSRF).'
      scheme: bearer
      bearerFormat: JWT